Integrate physics and collision with ECS

+++ Ukemochi-Engine/src/Ukemochi-Engine/Collision/Collision.cpp
@@ -71,7 +71,7 @@ namespace Ukemochi
 
 			// Step 2: Initialize and calculate the new velocity of Vb
 			firstTimeOfCollision = 0.0f;
-			float tLast = UME::g_FrameRateController.GetDeltaTime();
+			float tLast = static_cast<float>(UME::g_FrameRateController.GetDeltaTime());
 			//float tLast = (float)AEFrameRateControllerGetFrameTime(); // NEED ACCESS TO FRAME TIME!!
 			Vec2 vRel = { vel2.x - vel1.x, vel2.y - vel1.y };
